requirements.txt
langchain
langchain-community 
langchain-core
langchain-ollama
psycopg2
python-environ
